+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "axios": "^1.1.3",
     "vue": "^3.2.41",
+    "vue3-google-map": "^0.15.0",
     "vue3-smooth-scroll": "^0.8.1"
   },
   "devDependencies": {
